Implemented Texture Loading
Implemented basic texture loading. `LoadTextureFromFile()` creates a texture handle in `GameState::texture_handles` which maps the name of the texture (For example, "assets/gfx/grass.png" would be mapped as "grass.png") to an index into `GameState::textures`. Then the mapping in `GameState::texture_handles` can be used to look up a texture in `GameState::textures` by the filename. There are a few things which still need to be worked out.

- Currently you have to include the extension to the filename when loading, but this string will be mismatched from the string to do the lookup in `GameState::texture_handles`. It also is more typing when making metadata files that will come later.

I have left some commented out example code. This is because I don't have a solution to store binary data in GitHub and I don't want to pollute the repository.