Implemented the Entity Manager
The initial goal of the `EntityManager` to automatically handle the case when trying to perform some functionality but the component doesn't exist for that entity was misguided and instead should be handled by the Systems of the ECS rather than the `EntityManager`. However, the `EntityManager` still provided benefits by reducing the amount of code used to access components and use them. It also made creating entities simpler. I removed the functions for the Spawn Entity and Spawn Random Entity GUI because the random entity version isn't needed any more and the functionality of the Spawn Entity button got moved entirely to the `EntityManager`'s `CreateEntity()` function.