Added Basic Unit movement + supporting changes
Added basic hardcoded unit movement. No animations or transitions just getting things working. Properly follows range guideline set in the `BuildPath()` function. Added `movement_points` and `current_movement_points` to `cUnit` to use with `BuildPath()` to determine a unit's range. Added a display in the debug window of how many movement points are left. Movement is handled by selecting a unit and then right clicking on the map. R will reset the unit's movement_points to maximum. Modified `CreateEntity()` to return the index of the entity. This was used to improve behavior of the Spawn Entity button.