Implemented Initial Action System
The initial action system has been implemented. Actions are registered to an input method through `GameState::RegisterAction()` and will live in `GameState::action_map`. Then during the beginning of our main loop we check if any buttons were pressed or released and if they correspond to an `Action` that has been registered. We enqueue actions that occur into `GameState::action_queue`. Then in the game loop we pull the actions out of the queue and set the action boolean to true that is stored in `GameState` (this is an area that needs work). We then use the action booleans to execute the functionality that should occur this game loop.

Areas that could use improvement:
- Joystick actions, issue #51
- MouseWheel action
- Moving the input method adjustment functions out of `action.h/.c` and into a more general purpose raylib helper